+++ spritz/spritz-android/src/main/java/com/novoda/spritz/SpritzOnPageIdleListener.java
@@ -1,0 +1,7 @@
+package com.novoda.spritz;
+
+interface SpritzOnPageIdleListener {
+
+    void onPageIdle(int position);
+
+}
